seed fixed

+++ app/models/user.py
@@ -17,7 +17,7 @@ class User(db.Model, UserMixin):
     hashedPassword = db.Column(db.String(255), nullable=False)
     email = db.Column(db.String(255), nullable=False, unique=True)
     address=db.Column(db.String(100),nullable=False)
-    phone=db.Column(db.Integer(10),nullable=False)
+    phone=db.Column(db.Integer,nullable=False)
     created_at=db.Column(db.DateTime, default=datetime.now())
     updated_at=db.Column(db.DateTime, default=datetime.now())
